chore(clamav): Remove data volume
According to the [ClamAV documentation](https://docs.clamav.net/manual/Installing/Docker.html#image-selection-recommendations) you should either use the _base image and add persistent storage to the container or use the normal image without persistent storage.
There are reported issues with the database breaking for some time when
the pod is recreated while freshclam writes the database. This ensures
that a newly started pod is always functioning. Freshclam is run by the
container either way to retrieve updates (or fix a broken database) but
this way you will always start with a working one.

If persistent storage for the clamav pod is required the `clamav.extraVolumeMounts` or `clamav.extraVolumes` values in the helm chart can be used to achieve the same.

fixes #315

Signed-off-by: Adrian Nöthlich <git@promasu.tech>